requirements.txt
requests
python-dotenv
fastapi
uvicorn
playwright
httpx
